pokemon-backend/requirements.txt
anyio==3.6.2
ariadne==0.18.1
boto3==1.26.93
botocore==1.29.93
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
graphql-core==3.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
PyJWT==2.6.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.6
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
